refactor: 유저 생성 및 조회 로직 수정
