+++ wnp-server/src/main/java/com/wonnapark/wnpserver/domain/user/infrastructure/UserRepository.java
@@ -1,5 +1,6 @@
 package com.wonnapark.wnpserver.domain.user.infrastructure;
 
+import com.wonnapark.wnpserver.domain.user.OAuthProvider;
 import com.wonnapark.wnpserver.domain.user.User;
 import org.springframework.data.jpa.repository.JpaRepository;
 
@@ -7,5 +8,5 @@ import java.util.Optional;
 
 public interface UserRepository extends JpaRepository<User, Long> {
 
-    Optional<User> findByProviderId(Long providerId);
+    Optional<User> findByProviderIdAndPlatform(Long providerId, OAuthProvider platform);
 }
